extension/backend/requirements.txt
httpx
python-dotenv
pyelftools
pyserial
intelhex
typing-extensions
